requirements.txt
ciscoconfparse
pyyaml